requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
xlsxwriter==3.2.0

# PDF Generation
reportlab==4.0.9
matplotlib==3.8.2

# Email & Security
python-dotenv==1.0.0
secure-smtplib==0.1.1

# Data Visualization
plotly==5.18.0
seaborn==0.13.1

# Utilities
python-dateutil==2.8.2
Pillow==10.2.0